Better + descriptive nomenclature.  Refactored interface for acillary file models into abstracted module.
